# server/docs/openapi/spec.yaml
...
openapi: 3.1.1
info:
  title: Forms Portal API
  version: 0.1.0
  summary: API Documentation for the Forms Portal
  contact:
    name: OSDG
servers:
  - url: /api
    description: Production
...
security:
  - cookieAuth: []
paths:
...
  /forms:
    get:
      tags: [Forms]
      summary: List forms
      description: Retrieves forms accessible to the current user, with support for filtering, sorting, and pagination.
      operationId: listForms
      parameters:
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/PermissionRole'
        - name: sort
          in: query
          description: Sort criteria.
          schema:
            type: string
            enum: [modified, title]
            default: modified
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: A paginated list of forms.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Form'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    post:
      tags: [Forms]
      summary: Create form
      description: Creates a new form with the authenticated user as the owner.
      operationId: createForm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormCreate'
      responses:
        '201':
          description: Form created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /forms/{handle}/{slug}:
    parameters:
      - $ref: '#/components/parameters/handle'
      - $ref: '#/components/parameters/slug'
    get:
      tags: [Forms]
      summary: Resolve form
      description: Resolves a user-friendly URL path (`/handle/slug`) to a full form object.
      operationId: resolveForm
      responses:
        '200':
          description: Form resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        '403':
          $ref: '#/components/responses/Forbidden'
  /forms/{formId}:
    parameters:
      - $ref: '#/components/parameters/formId'
    get:
      tags: [Forms]
      summary: Get form by ID
      description: Retrieves a form's complete definition. Requires VIEW permission or higher.
      operationId: getForm
      responses:
        '200':
          description: Form details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    patch:
      tags: [Forms]
      summary: Update form
      description: Updates a form's definition. Requires EDIT permission or higher.
      operationId: updateForm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormUpdate'
      responses:
        '200':
          description: Form updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Forms]
      summary: Delete form
      description: Permanently deletes a form and all associated data. This action cannot be undone. Requires MANAGE permission.
      operationId: deleteForm
      responses:
        '204':
          description: Form deleted successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /forms/{formId}/permissions:
    parameters:
      - $ref: '#/components/parameters/formId'
    get:
      tags: [Permissions]
      summary: List form permissions
      description: Retrieves all active permission grants for a form. Requires MANAGE permission.
      operationId: listPermissions
      responses:
        '200':
          description: List of permission grants.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Permissions]
      summary: Grant permission
      description: Grants a role on a form to a user or a group. Requires MANAGE permission.
      operationId: grantPermission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionCreate'
      responses:
        '201':
          description: Permission granted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /forms/{formId}/permissions/{permissionId}:
    parameters:
      - $ref: '#/components/parameters/formId'
      - $ref: '#/components/parameters/permissionId'
    delete:
      tags: [Permissions]
      summary: Revoke permission
      description: Removes a permission grant from a user or group. Requires MANAGE permission.
      operationId: revokePermission
      responses:
        '204':
          description: Permission revoked successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cookieAuth:
      type: cookie
      name: session
  parameters:
    handle:
      name: handle
      in: path
      required: true
      schema:
        type: string
    slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
    formId:
      name: formId
      in: path
      required: true
      schema:
        type: string
        format: ulid
...
    permissionId:
      name: permissionId
      in: path
      required: true
      schema:
        type: string
        format: ulid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
    PermissionRole:
      type: string
      enum: [view, respond, comment, analyze, edit, manage]
...
    Form:
      type: object
      required:
        - id
        - owner
        - title
        - slug
        - structure
        - live
      properties:
        id:
          type: string
          format: ulid
        owner:
          type: string
          format: ulid
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        structure:
          type: string
          format: kdl
        live:
          type: boolean
          default: false
        opens:
          type: string
          format: date-time
          nullable: true
        closes:
          type: string
          format: date-time
          nullable: true
        anonymous:
          type: boolean
          default: false
        max_responses:
          type: integer
          nullable: true
        individual_limit:
          type: integer
          minimum: 1
          default: 1
        editable_responses:
          type: boolean
          default: false
    FormCreate:
      type: object
      required:
        - title
        - slug
        - structure
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        structure:
          type: string
          format: kdl
        live:
          type: boolean
          default: false
        opens:
          type: string
          format: date-time
          nullable: true
        closes:
          type: string
          format: date-time
          nullable: true
        anonymous:
          type: boolean
        max_responses:
          type: integer
          nullable: true
        individual_limit:
          type: integer
          minimum: 1
          default: 1
        editable_responses:
          type: boolean
          default: false
    FormUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        structure:
          type: string
          format: kdl
        live:
          type: boolean
        opens:
          type: string
          format: date-time
          nullable: true
        closes:
          type: string
          format: date-time
          nullable: true
        anonymous:
          type: boolean
        max_responses:
          type: integer
          nullable: true
        individual_limit:
          type: integer
          minimum: 1
        editable_responses:
          type: boolean
...
    Permission:
      type: object
      required:
        - id
        - form
        - role
      properties:
        id:
          type: string
          format: ulid
        form:
          type: string
          format: ulid
        role:
          $ref: '#/components/schemas/PermissionRole'
        user:
          type: string
          format: ulid
          nullable: true
        group:
          type: string
          format: ulid
          nullable: true
    PermissionCreate:
      type: object
      required:
        - role
      oneOf:
        - required: [user]
        - required: [group]
      properties:
        role:
          $ref: '#/components/schemas/PermissionRole'
        user:
          type: string
          format: email
        group:
          type: string
          format: ulid
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    ValidationError:
      type: object
      required:
        - code
        - message
        - errors
      properties:
        code:
          type: string
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Pagination:
      type: object
      required:
        - offset
        - limit
        - total
      properties:
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer